refactor(examples): split domain code into .go files, rename test files, add READMEs
