+++ examples/auth/suite_ext_test.go
