+++ examples/cart/README.md
@@ -1,0 +1,14 @@
+# cart — Shopping Cart Operations
+
+Demonstrates basic assertions and error handling using a shopping cart with
+add, remove, discount, and checkout operations.
+
+## Structure
+
+- **cart.go** — Cart with catalog pricing, quantity tracking, and discount logic
+- **suite_test.go** — `ShoppingCartTestSuite` with 4 test methods
+- **suite_ext_test.go** — `ShoppingCartTestSuite` (external package variant)
+
+## Features
+
+`Equal` · `Contains` · `Empty` · `InDelta` · `GreaterOrEqual` · `ErrorIs` · `NoError` · `Len` · `BeforeEach`
