+++ examples/repository/inventory/README.md
@@ -1,0 +1,13 @@
+# inventory — Comparison Assertions
+
+Demonstrates the full set of ordered comparison assertions using stock level
+management with reserve and restock operations.
+
+## Structure
+
+- **inventory.go** — Stock level with available quantity, reserve, and restock
+- **suite_test.go** — `InventoryTestSuite` with `BeforeEach`/`AfterEach`
+
+## Features
+
+`Less` · `Greater` · `LessOrEqual` · `GreaterOrEqual` · `NotEqual` · `Equal` · `True` · `False` · `BeforeEach` · `AfterEach`
