+++ examples/repository/inventory/suite_test.go
@@ -2,32 +2,6 @@ package inventory
 
 import "github.com/mvrahden/go-test/pkg/gotest"
 
-// --- Domain types ---
-
-type StockLevel struct {
-	SKU      string
-	Quantity int
-	Reserved int
-}
-
-func (s *StockLevel) Available() int {
-	return s.Quantity - s.Reserved
-}
-
-func (s *StockLevel) Reserve(qty int) bool {
-	if qty > s.Available() {
-		return false
-	}
-	s.Reserved += qty
-	return true
-}
-
-func (s *StockLevel) Restock(qty int) {
-	s.Quantity += qty
-}
-
-// --- Test Suite ---
-
 type InventoryTestSuite struct {
 	stock *StockLevel
 }
