+++ examples/repository/suite_ext_test.go
