+++ examples/search/suite_ext_test.go
